# docs/api/openapi/llm-processor-openapi.yaml
openapi: 3.0.3
info:
  title: Nephoran LLM Processor Service API
  description: |
    Enterprise-grade LLM processing service for the Nephoran Intent Operator.
    This service processes natural language intents and converts them to structured network operations.
  version: 2.0.0
  contact:
    name: Nephoran Development Team
    email: support@nephoran.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://llm-processor:8080
    description: Default service endpoint
  - url: http://localhost:8080
    description: Local development endpoint
  - url: https://llm-processor.nephoran.com
    description: Production endpoint
...
security:
  - BearerAuth: []
  - OAuth2: [operator]
paths:
  /process:
    post:
      summary: Process natural language intent
      description: Convert natural language intents into structured network operations
      tags:
        - Processing
      operationId: processIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessIntentRequest'
      responses:
        '200':
          description: Intent processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessIntentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /auth/login/{provider}
          tokenUrl: /auth/callback/{provider}
          scopes:
            operator: Process intents and view status
            admin: Full administrative access
  schemas:
    ProcessIntentRequest:
      type: object
      required:
        - intent
      properties:
        intent:
          type: string
          description: Natural language intent to process
          example: "Deploy AMF with 3 replicas for network slice eMBB with high throughput requirements"
        metadata:
          type: object
          description: Additional metadata for processing
          properties:
            namespace:
              type: string
              example: telecom-core
            priority:
              type: string
              enum: [low, medium, high, critical]
              example: high
            user_id:
              type: string
              example: operator-001
        config:
          type: object
          description: Processing configuration
          properties:
            model:
              type: string
              default: gpt-4o-mini
              example: gpt-4o-mini
            max_tokens:
              type: integer
              default: 2048
              example: 2048
            enable_rag:
              type: boolean
              default: true
              example: true
    ProcessIntentResponse:
      type: object
      properties:
        result:
          type: string
          description: Processed result
          example: "Generated Kubernetes YAML configuration for AMF deployment..."
        status:
          type: string
          enum: [success, error]
          example: success
        processing_time:
          type: string
          example: "2.345s"
        request_id:
          type: string
          example: "req-1641024000123456789"
        metadata:
          type: object
          properties:
            tokens_used:
              type: integer
              example: 1456
            model_used:
              type: string
              example: gpt-4o-mini
            confidence_score:
              type: number
              format: float
              example: 0.94
            rag_context_used:
              type: boolean
              example: true
            context_length:
              type: integer
              example: 4096
        structured_output:
          type: object
          description: Structured representation of the processed intent
          properties:
            action:
              type: string
              example: deploy
            resources:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    example: Deployment
                  name:
                    type: string
                    example: amf-deployment
                  replicas:
                    type: integer
                    example: 3
                  namespace:
                    type: string
                    example: telecom-core
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INTENT_PROCESSING_FAILED
            type:
              type: string
              example: processing_error
            message:
              type: string
              example: Failed to process network intent due to invalid parameters
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
              example: "2025-01-30T10:30:15Z"
            request_id:
              type: string
              example: req-12345
            correlation_id:
              type: string
              example: corr-67890
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per minute
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when rate limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Service unavailable - Circuit breaker open or dependencies unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'